OFS-88: fixing folder name qgl->gql
